Move patch related codes from models to pull service
